fix korean encode Error and refactor codes

+++ front/FileChecker.py
@@ -2,7 +2,8 @@ from File import File, FileList
 import asyncio
 import os, re
 import uuid
-from Error import CustomError
+from API import API
+from Error import AlreadyChecked, NeedToUpdated
 try:
     from watchdog.observers import Observer
     from watchdog.events import RegexMatchingEventHandler
@@ -11,68 +12,70 @@ except ModuleNotFoundError as e:
     os.system("pip install watchdog")
 
 class FileChecker(RegexMatchingEventHandler):
-    def __init__(self, _target, _logger, _api):
+    def __init__(self, _target, _logger, _config):
         super().__init__(None, [".*\.py",".*\.pyc",".*\.ini", ".*\.DS\_Store", ".*thumbs\.db"], False, False)
         self.target = _target
         self.filelist = FileList(self.target)
         self.logger = _logger
-        self.api = _api
-        
+        self.config = _config
+        self.api = API(self.config, self.logger)
+        self.observer = self.setObserver(self.target)
         # initial file sync
         self.syncToServer(self.findToUpdate())
-        
-        self.observer = self.setObserver(_target)
 
     def findToUpdate(self):
+        self.logger.print_log("GETTING FILE DATAS FROM SERVER")
         need_update = asyncio.run(self.api.getFileList())
-        print(need_update)
+        self.logger.print_log("ARRIVED FILE DATAS  TO  CLIENT")
+        self.logger.print_file_list(need_update)
+        self.logger.print_log("CHECKING FILE TO UPDATE...")
+        self.logger.print_log("--------------------------")
         for (path, dir, file) in os.walk(self.target):
             for f in file:
+                if f == ".DS_Store":
+                    continue
                 f_loc = self.filelist.append("%s/%s" % (path, f))
-                # print(f_loc)
                 try:
                     for f_ser in need_update:
                         if f_loc.sync_path == f_ser["path"]:
                             if f_loc.md5 == f_ser["md5"]:
                                 f_loc.id = uuid.UUID(f_ser["id"])
                                 need_update.remove(f_ser)
-                                raise CustomError("Already checked")
+                                raise AlreadyChecked("NO NEED TO UPDATE %s" % f_loc.name)
                             else:
-                                print(f_loc.name, "Delete Content modified file")
                                 os.remove(f_loc.real_path)
                                 self.filelist.pop()
-                                raise CustomError("Already checked")
+                                raise AlreadyChecked("NEED TO UPDATE %s" %f_loc.name)
                             break
-                    raise NotImplementedError("Delete path modified file")
-                except NotImplementedError as e:
-                    print(f_loc.name, e)
+                    raise NeedToUpdated("NEED TO UPDATE %s" % f_loc.name)
+                except NeedToUpdated as e:
+                    self.logger.print_log(str(e))
                     os.remove(f_loc.real_path)
                     self.filelist.pop()
-                except CustomError as e:
+                except AlreadyChecked as e:
+                    self.logger.print_log(str(e))
                     continue
+            if os.path.exists("%s/.DS_Store" % path):
+                os.remove("%s/.DS_Store" % path)
                 
         for (path, dir, file) in os.walk(self.target):
             for d in dir:
                 try:
                     os.rmdir("%s/%s" % (path, d))
                 except:
-                    print("Directory is not Empty :", d)
                     continue
+        self.logger.print_log("--------------------------")
+        self.logger.print_log("ALL FILE CHECKED")
         return need_update
     
     def syncToServer(self, need_update):
-        # 파일 데이터 생성
+        self.logger.print_log("SYNCHRONIZING TO SERVER")
         for file in need_update:
-            ############
-            # 파일 다운로드
-            ############
             asyncio.run(self.api.downloadFile(file["id"]))
-            
-            # 다운로드 한 뒤에 append해야 size랑 md5 계산 가능
-            f = self.filelist.append("%s%s" % (self.target, file["path"][:5]))
+            f = self.filelist.append("%s%s" % (self.target, file["path"][4:]))
             f.id = uuid.UUID(file["id"])
-
-        # print(self.filelist)
+        self.logger.print_log("SYNCHRONIZATION COMPLETE")
+        self.observer.start()
     
     def setObserver(self, target):
         observer = Observer()
@@ -82,26 +85,24 @@ class FileChecker(RegexMatchingEventHandler):
     def on_created(self, event):
         if event.is_directory:
             return
+        
+        if self.filelist.search(event.src_path):
+            return
         f = self.filelist.append(event.src_path)
-        # 서버에 파일 정보 등록
+        self.logger.print_log("FILE '%s' CREATED AT '%s'" % (f.name, f.dir))
         asyncio.run(self.api.createFile(f))
-        
-        ###########
-        # 파일 업로드
-        ###########
         asyncio.run(self.api.uploadFile(f))
-        
-        self.logger.print_log("File '%s' created at %s" % (f.name, f.dir))
+        # socket file created
         
     def on_moved(self, event):
         f = self.filelist.move(event.src_path, event.dest_path)
-        asyncio.run(self.api.modifyFile(f))
-        asyncio.run(self.api.uploadFile(f))
         if f == -1:
-            self.logger.print_log("MOVE ERROR :", event.src_path, "to", event.dest_path)
+            self.logger.print_log("MOVE ERROR :", event.src_path, "TO", event.dest_path)
             return
         else:
-            self.logger.print_log("File '%s' moved \n\t\t\tfrom %s \n\t\t\tto   %s" % (f.name, re.sub(self.target, "Root", event.src_path), re.sub(self.target, "Root", event.dest_path)))
+            self.logger.print_log("FILE '%s' MOVED \n%sFROM %s \n%sTO%s" % (f.name, " "*21, re.sub(self.target, "Root", event.src_path), " "*21, re.sub(self.target, "Root", event.dest_path)))
+            asyncio.run(self.api.modifyFile(f))
+            # socket file moved
     
     def on_deleted(self, event):
         if event.is_directory:
@@ -111,18 +112,20 @@ class FileChecker(RegexMatchingEventHandler):
                 return
             else:
                 for f in fs:
+                    self.logger.print_log("FILE '%s' DELETED FROM '%s'" % (f.name, f.dir))
                     asyncio.run(self.api.deleteFile(f))
-                    self.logger.print_log("File '%s' deleted at %s" % (f.name, f.dir))
                     del f
+                # socket Delete fs lists
         else:
             f = self.filelist.del_file(event.src_path)
             if f == -1:
-                self.logger.print_log("DELETE FILE ERROR :", event.src_path)
+                self.logger.print_log("DELETE FILE ERROR : %s" % event.src_path)
                 return
             else:
+                self.logger.print_log("FILE '%s' DELETED FROM %s" % (f.name, f.dir))
                 asyncio.run(self.api.deleteFile(f))
-                self.logger.print_log("File '%s' deleted at %s" % (f.name, f.dir))
                 del f
+                # socket Delete f
         
     def on_modified(self, event):
         if event.is_directory:
